move demo data into add-on data package
