--- vaadin-maps-leaflet-flow-demo/src/main/java/software/xdev/vaadin/maps/leaflet/flow/demo/data/repository/MarkerRepository.java
@@ -1,8 +1,0 @@
-package software.xdev.vaadin.maps.leaflet.flow.demo.data.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.entity.Marker;
-
-
-public interface MarkerRepository extends JpaRepository<Marker, Long> {
-}
