+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/data/entity/Marker.java
@@ -1,15 +1,14 @@
-package software.xdev.vaadin.maps.leaflet.flow.demo.data.entity;
+package software.xdev.vaadin.maps.leaflet.flow.data.entity;
 
 import java.util.List;
 
 import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 
-import software.xdev.vaadin.maps.leaflet.flow.data.LPoint;
-
 
 @Entity
-public class Marker extends AbstractEntity {
+public class Marker extends AbstractEntity
+{
 	
 	@ElementCollection
 	private List<Double> point;
