+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/data/entity/Polyline.java
@@ -1,17 +1,16 @@
-package software.xdev.vaadin.maps.leaflet.flow.demo.data.entity;
+package software.xdev.vaadin.maps.leaflet.flow.data.entity;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.validation.constraints.NotNull;
-
 import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import jakarta.persistence.OneToMany;
-import software.xdev.vaadin.maps.leaflet.flow.data.LPoint;
+
 
 @Entity
-public class Polyline extends AbstractEntity {
+public class Polyline extends AbstractEntity
+{
 	
 
 	// private List<LPoint> points = new ArrayList<>();
