+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/data/entity/Rectangle.java
@@ -1,4 +1,4 @@
-package software.xdev.vaadin.maps.leaflet.flow.demo.data.entity;
+package software.xdev.vaadin.maps.leaflet.flow.data.entity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,10 +7,10 @@ import java.util.List;
 import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 
-import software.xdev.vaadin.maps.leaflet.flow.data.LPoint;
 
 @Entity
-public class Rectangle extends AbstractEntity {
+public class Rectangle extends AbstractEntity
+{
 	
 	@ElementCollection
 	private List<Double> points = new ArrayList<>();
