+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/data/repository/MarkerRepository.java
@@ -1,0 +1,9 @@
+package software.xdev.vaadin.maps.leaflet.flow.data.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import software.xdev.vaadin.maps.leaflet.flow.data.entity.Marker;
+
+
+public interface MarkerRepository extends JpaRepository<Marker, Long> {
+}
