+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/data/repository/PolylineRepository.java
@@ -1,7 +1,8 @@
-package software.xdev.vaadin.maps.leaflet.flow.demo.data.repository;
+package software.xdev.vaadin.maps.leaflet.flow.data.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.entity.Polyline;
+
+import software.xdev.vaadin.maps.leaflet.flow.data.entity.Polyline;
 
 
 public interface PolylineRepository extends JpaRepository<Polyline, Long>
