+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/data/service/DbService.java
@@ -1,15 +1,15 @@
-package software.xdev.vaadin.maps.leaflet.flow.demo.data.service;
+package software.xdev.vaadin.maps.leaflet.flow.data.service;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.entity.Marker;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.entity.Polyline;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.entity.Rectangle;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.repository.MarkerRepository;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.repository.PolylineRepository;
-import software.xdev.vaadin.maps.leaflet.flow.demo.data.repository.RectangleRepository;
+import software.xdev.vaadin.maps.leaflet.flow.data.entity.Marker;
+import software.xdev.vaadin.maps.leaflet.flow.data.entity.Polyline;
+import software.xdev.vaadin.maps.leaflet.flow.data.repository.MarkerRepository;
+import software.xdev.vaadin.maps.leaflet.flow.data.repository.PolylineRepository;
+import software.xdev.vaadin.maps.leaflet.flow.data.entity.Rectangle;
+import software.xdev.vaadin.maps.leaflet.flow.data.repository.RectangleRepository;
 
 
 @Service
@@ -32,6 +32,8 @@ public class DbService {
 		markerRepository.save(marker);
 	}
 	
+	public void deleteMarker(Marker marker) { markerRepository.delete(marker); }
+	
 	public void savePolyline(Polyline polyline) {
 		if (polyline == null) {
 			System.err.println("Error, Polyline is null");
@@ -40,6 +42,8 @@ public class DbService {
 		polylineRepository.save(polyline);
 	}
 	
+	public void deletePolyline(Polyline polyline) { polylineRepository.delete(polyline); }
+	
 	public void saveRectangle(Rectangle rectangle) {
 		if (rectangle == null) {
 			System.err.println("Error, Rectangle is null");
@@ -47,10 +51,11 @@ public class DbService {
 		}
 		rectangleRepository.save(rectangle);
 	}
+	
+	public void deleteRectangle(Rectangle rectangle) { rectangleRepository.delete(rectangle); }
 
 	
 	public List<Marker> findAllMarkers() { return markerRepository.findAll(); }
 	public List<Polyline> findAllPolylines() { return polylineRepository.findAll(); }
 	public List<Rectangle> findAllRectangles() { return rectangleRepository.findAll(); }
-	
 }
